Created operations directory
Created operations directory. Removed app.route functions from web_service.py
